xde_steady, dde2ode, compute_terms, parse_deout, Rainfall, Temperature, Relative_Humidity
xde_steady was added to compute the steady states of models

Since xde_steady works on ode, we also added the dde2ode family of functions

Weather was configured to call Rainfall, Temperature, Relative_Humidity

parse_deout was changed to return variable names attached to either XH, MYZ, or L to avoid conflicts if two models used the same variable name.

the compute_terrms -- an S3 family of functions -- was added. It is now called from parse_deout. xde_setup was modified to add the term $compute with a class assigned to dispatch the S3 function